requirements.txt
asgiref==3.9.0
Brotli==1.1.0
cffi==1.17.1
cssselect2==0.8.0
diff-match-patch==20241021
dj-database-url==3.0.1
Django==5.2.4
django-import-export==4.3.8
djangorestframework==3.16.0
fonttools==4.58.5
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-decouple==3.8
sqlparse==0.5.3
tablib==3.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.1
tzdata==2025.2
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.9.0
zopfli==0.2.3.post1
